+++ otterdog/operations/push_config.py
@@ -6,6 +6,8 @@
 #  SPDX-License-Identifier: EPL-2.0
 #  *******************************************************************************
 
+import filecmp
+
 import aiofiles
 import aiofiles.ospath
 from git import InvalidGitRepositoryError, Repo
@@ -13,9 +15,10 @@ from git.config import GitConfigParser
 
 from otterdog.config import OrganizationConfig
 from otterdog.providers.github import GitHubProvider
-from otterdog.utils import style
+from otterdog.utils import get_approval, style
 
 from . import Operation
+from .local_plan import LocalPlanOperation
 
 
 class PushOperation(Operation):
@@ -23,9 +26,14 @@ class PushOperation(Operation):
     Pushes a local configuration of an organization to its meta-data repository.
     """
 
-    def __init__(self, push_message: str | None):
+    def __init__(self, show_diff: bool, push_message: str | None):
         super().__init__()
+        self._show_diff = show_diff
         self._push_message = push_message
+
+    @property
+    def show_diff(self) -> bool:
+        return self._show_diff
 
     @property
     def push_message(self) -> str | None:
@@ -86,21 +94,24 @@ class PushOperation(Operation):
 
             async with GitHubProvider(credentials) as provider:
                 try:
+                    perform_update = True
                     updated_files = []
-                    updated = False
 
-                    if await provider.update_content(
-                        org_config.github_id,
-                        org_config.config_repo,
-                        f"otterdog/{github_id}.jsonnet",
-                        content,
-                        None,
-                        self.push_message,
-                        author_name,
-                        author_email,
-                    ):
-                        updated_files.append(f"otterdog/{github_id}.jsonnet")
-                        updated = True
+                    if self.show_diff:
+                        perform_update = await self._display_diff(org_config, provider)
+
+                    if perform_update:
+                        if await provider.update_content(
+                            org_config.github_id,
+                            org_config.config_repo,
+                            f"otterdog/{github_id}.jsonnet",
+                            content,
+                            None,
+                            self.push_message,
+                            author_name,
+                            author_email,
+                        ):
+                            updated_files.append(f"otterdog/{github_id}.jsonnet")
 
                 except RuntimeError as e:
                     self.printer.print_error(
@@ -108,7 +119,7 @@ class PushOperation(Operation):
                     )
                     return 1
 
-            if updated:
+            if len(updated_files) > 0:
                 self.printer.println(
                     f"organization definition pushed to repo '{org_config.github_id}/{org_config.config_repo}': "
                 )
@@ -120,3 +131,48 @@ class PushOperation(Operation):
             return 0
         finally:
             self.printer.level_down()
+
+    async def _display_diff(self, org_config: OrganizationConfig, provider: GitHubProvider) -> bool:
+        rest_api = provider.rest_api
+        github_id = org_config.github_id
+
+        repo_data = await rest_api.repo.get_repo_data(org_config.github_id, org_config.config_repo)
+        default_branch = repo_data["default_branch"]
+
+        current_definition = await provider.get_content(
+            github_id,
+            org_config.config_repo,
+            f"otterdog/{github_id}.jsonnet",
+            default_branch,
+        )
+
+        current_config_file = org_config.jsonnet_config.org_config_file + "-BASE"
+        async with aiofiles.open(current_config_file, "w") as file:
+            await file.write(current_definition)
+
+        if filecmp.cmp(current_config_file, org_config.jsonnet_config.org_config_file):
+            return False
+
+        self.printer.println("The following changes compared to the current configuration exist locally:")
+        self.printer.println()
+        self.printer.level_up()
+
+        try:
+            operation = LocalPlanOperation("-BASE", False, False, "")
+            operation.init(self.config, self.printer)
+            valid_config = await operation.generate_diff(org_config)
+        finally:
+            self.printer.level_down()
+
+        if valid_config != 0:
+            return False
+
+        self.printer.println()
+        self.printer.println("Do you want to push these changes? " "(Only 'yes' or 'y' will be accepted as approval)\n")
+
+        self.printer.print(f"{style('Enter a value', bright=True)}: ")
+        if not get_approval():
+            self.printer.println("\npush cancelled.")
+            return False
+
+        return True